Remove dependency installation.
Installing and configuring all dependency applications is going to be prohibitively complicated. Installing Git is fairly easy (and almost never actually necessary) so we'll leave that in the install script for now. The rest of the scripts will only bootstrap the dotfiles. Any dependenct applications can be installed as needed.