Convert to formatter.nvim and nvim-lint.
This offers less functionality than null-ls but null-ls has a bug with phpcbf where the buffer content gets replaced with the output of the command. Since it's archived, that won't get fixed until someone picks it up. In the meantime, formatter.nvim and nvim-lint give me what I need for PHP at least. We can find a more complex solution if it becomes necessary in the future.